+++ lib/api-key/mod.ts
@@ -4,4 +4,5 @@ export * from "./generators.ts";
 export * from "./etag.ts";
 export * from "./service.ts";
 export * from "./memory-repository.ts";
+export * from "./kv-repository.ts";
 export * from "./http.ts";